refactor(calendar): move calendar doctypes from Mail to Calendar module
Relocate the 10 calendar-related doctypes (Calendar, Calendar Event, Calendar
Exchange, Calendar Rights, Event Alert, Event Calendar, Event Link, Event
Location, Event Notification, Event Participant) into the Calendar module and
repoint all dependent code.

- git mv the doctype folders to suite/calendar/doctype and set module=Calendar.
- Update Python imports, hooks (has_permission, permission_query_conditions,
  scheduler_events), frontend API URLs, and desk list scripts.

The Import/Export UI in Mail Settings and the JMAP calendar services stay in
Mail for now (only the mechanical fetch_calendars URL was repointed).
