+++ frontend/src/apps/mail/components/Settings/CalendarImportSettings.vue
@@ -92,7 +92,7 @@ const onFileSelected = async (event: Event) => {
 }
 
 const calendars = createResource({
-	url: 'suite.mail.doctype.calendar.calendar.fetch_calendars',
+	url: 'suite.calendar.doctype.calendar.calendar.fetch_calendars',
 	auto: true,
 	makeParams: () => ({ account: accountId, limit: 100 }),
 	onSuccess: (data: { id: string }[]) => {
