+++ suite/calendar/doctype/calendar/calendar.json
@@ -252,7 +252,7 @@
  ],
  "modified": "2026-06-30 10:19:37.522414",
  "modified_by": "Administrator",
- "module": "Mail",
+ "module": "Calendar",
  "name": "Calendar",
  "owner": "Administrator",
  "permissions": [
