+++ suite/calendar/doctype/calendar_event/calendar_event.js
