+++ suite/calendar/doctype/calendar_event/invite_templates.py
