+++ suite/calendar/doctype/calendar_event/test_calendar_event.py
