+++ suite/calendar/doctype/calendar_rights/calendar_rights.json
@@ -101,7 +101,7 @@
  "links": [],
  "modified": "2026-01-15 21:38:35.292181",
  "modified_by": "Administrator",
- "module": "Mail",
+ "module": "Calendar",
  "name": "Calendar Rights",
  "owner": "Administrator",
  "permissions": [],
