+++ suite/calendar/doctype/calendar_rights/calendar_rights.py
