+++ suite/calendar/doctype/event_alert/event_alert.json
@@ -85,7 +85,7 @@
  "links": [],
  "modified": "2026-01-23 15:00:14.568635",
  "modified_by": "Administrator",
- "module": "Mail",
+ "module": "Calendar",
  "name": "Event Alert",
  "owner": "Administrator",
  "permissions": [],
