+++ suite/calendar/doctype/event_calendar/event_calendar.json
@@ -58,7 +58,7 @@
  "links": [],
  "modified": "2026-01-23 15:02:52.160541",
  "modified_by": "Administrator",
- "module": "Mail",
+ "module": "Calendar",
  "name": "Event Calendar",
  "owner": "Administrator",
  "permissions": [],
