+++ suite/calendar/doctype/event_link/event_link.json
@@ -52,7 +52,7 @@
  "links": [],
  "modified": "2026-01-23 14:56:53.652926",
  "modified_by": "Administrator",
- "module": "Mail",
+ "module": "Calendar",
  "name": "Event Link",
  "owner": "Administrator",
  "permissions": [],
