+++ suite/calendar/doctype/event_location/event_location.json
@@ -43,7 +43,7 @@
  "links": [],
  "modified": "2026-01-23 15:01:21.498580",
  "modified_by": "Administrator",
- "module": "Mail",
+ "module": "Calendar",
  "name": "Event Location",
  "owner": "Administrator",
  "permissions": [],
