+++ suite/calendar/doctype/event_location/event_location.py
