+++ suite/calendar/doctype/event_notification/event_notification.js
