+++ suite/calendar/doctype/event_participant/event_participant.json
@@ -126,7 +126,7 @@
  "links": [],
  "modified": "2026-03-25 13:04:04.214533",
  "modified_by": "Administrator",
- "module": "Mail",
+ "module": "Calendar",
  "name": "Event Participant",
  "owner": "Administrator",
  "permissions": [],
