+++ suite/hooks.py
@@ -131,7 +131,7 @@ permission_query_conditions = {
 	# mail
 	"JMAP Account": "suite.mail.doctype.jmap_account.jmap_account.get_permission_query_condition",
 	"Blocked Email Address": "suite.mail.doctype.blocked_email_address.blocked_email_address.get_permission_query_condition",
-	"Calendar Exchange": "suite.mail.doctype.calendar_exchange.calendar_exchange.get_permission_query_condition",
+	"Calendar Exchange": "suite.calendar.doctype.calendar_exchange.calendar_exchange.get_permission_query_condition",
 	"Contacts Exchange": "suite.mail.doctype.contacts_exchange.contacts_exchange.get_permission_query_condition",
 	"Junk Email Address": "suite.mail.doctype.junk_email_address.junk_email_address.get_permission_query_condition",
 	"Mail Exchange": "suite.mail.doctype.mail_exchange.mail_exchange.get_permission_query_condition",
@@ -159,12 +159,12 @@ has_permission = {
 	"JMAP Account": "suite.mail.doctype.jmap_account.jmap_account.has_permission",
 	"Address Book": "suite.mail.doctype.address_book.address_book.has_permission",
 	"Blocked Email Address": "suite.mail.doctype.blocked_email_address.blocked_email_address.has_permission",
-	"Calendar": "suite.mail.doctype.calendar.calendar.has_permission",
-	"Calendar Event": "suite.mail.doctype.calendar_event.calendar_event.has_permission",
-	"Calendar Exchange": "suite.mail.doctype.calendar_exchange.calendar_exchange.has_permission",
+	"Calendar": "suite.calendar.doctype.calendar.calendar.has_permission",
+	"Calendar Event": "suite.calendar.doctype.calendar_event.calendar_event.has_permission",
+	"Calendar Exchange": "suite.calendar.doctype.calendar_exchange.calendar_exchange.has_permission",
 	"Contact Card": "suite.mail.doctype.contact_card.contact_card.has_permission",
 	"Contacts Exchange": "suite.mail.doctype.contacts_exchange.contacts_exchange.has_permission",
-	"Event Notification": "suite.mail.doctype.event_notification.event_notification.has_permission",
+	"Event Notification": "suite.calendar.doctype.event_notification.event_notification.has_permission",
 	"Identity": "suite.mail.doctype.identity.identity.has_permission",
 	"Junk Email Address": "suite.mail.doctype.junk_email_address.junk_email_address.has_permission",
 	"Mail Exchange": "suite.mail.doctype.mail_exchange.mail_exchange.has_permission",
@@ -253,13 +253,13 @@ scheduler_events = {
 		"suite.mail.doctype.jmap_account.jmap_account.delete_orphaned_jmap_accounts",
 		"suite.mail.doctype.mail_exchange.mail_exchange.clean_import_export_directories",
 		"suite.mail.doctype.push_subscription.push_subscription.renew_expiring_push_subscriptions",
-		"suite.mail.doctype.calendar_exchange.calendar_exchange.clean_calendar_import_export_directories",
+		"suite.calendar.doctype.calendar_exchange.calendar_exchange.clean_calendar_import_export_directories",
 		"suite.mail.doctype.contacts_exchange.contacts_exchange.clean_contacts_import_export_directories",
 	],
 	"hourly": [
 		# mail
 		"suite.mail.doctype.mail_exchange.mail_exchange.retry_stuck_mail_exchanges",
-		"suite.mail.doctype.calendar_exchange.calendar_exchange.retry_stuck_calendar_exchanges",
+		"suite.calendar.doctype.calendar_exchange.calendar_exchange.retry_stuck_calendar_exchanges",
 		"suite.mail.doctype.contacts_exchange.contacts_exchange.retry_stuck_contacts_exchanges",
 	],
 	"hourly_long": [
